ai/openai/requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
pydantic==2.5.0
openai==1.35.7
python-dotenv==1.0.0
httpx==0.27.0